Use closure instead of stream to stop iteration.
`IterativeScheduler` relies on the user to decide when to stop
iterating.  This way we can easily implement different iterative
behaviors, like fixed points or a fixed number of iterations.
Initially, the termination condition was read from a `Stream<_,bool>`,
returned by the nested circuit constructor.  We change this to a simpler
design where the circuit constructor returns `Fn()->bool` instead.