Operators over streams: Z1, integrate.
Versions of z^-1 and integrator that work on streams rather than
individual values.  Mathematically, these are identical to regular `Z1`
and `integrate` operators, but implementation-wise these operators
consume the inputs value-by-value as the (conceptually infinite) input
stream gets generated.

This commit also introduces an adapter mechanism that makes it easy to
wrap values in streams in shared references (`Rc`, `Arc`, `Intern`,
etc).  Adapter operators encapsulate regular operators, which do not need
to know about such references, and take care of unwrapping inputs and
wrapping outputs.  This mechanism allows shared values to be stored
(read-only) in multiple operators without memory bloat.  We introduce
this capability now since it will be useful in  combination with the new
`integrate_nested` operator.  Consider the following circuit, that will
arise in the implementation of nested incremental join:

```
       integrate_nested
      ┌───────────────┐
      │               │
      │     ┌───┐     │
      │     │   │     │
──────┼─────┤ + ├───┬─┼─ Z ─────►
      │     │   │   │ │
      │     └───┘   │ │
      │       ▲     │ │
      │       │     │ │
      │ ZNested ◄───┘ │
      │               │
      └───────────────┘
```

Here the `ZNested` that is part of the nested integrator stores a vector
of Z-sets, one per nested timestamp.  The "regular" `Z` operator
(\lift(Z)) contains a copy of the second-last element of the vector.