tracing: add `SubcircuitComplete` event type.
This commit sets things up for adding support for regions in the next
commit.  A region is a logical grouping of operators for debugging and
visualization purposes.  Regions only exist inside the circuit monitor
and won't affect circuit execution in any way.  When the client code
declares a region, this region will belong to the current circuit that
is being constructed at the moment, i.e., the circuit passed to the
innermost `Root::build` or `Circuit::iterate` closure.

In order to enable the monitor to track the current circuit, we bracket
subcircuit construction in a pair of `Subcircuit`/`SubcircuitComplete`
events, which push and pop a subcircuit on the stack respectively.  The
latter event is new.