Circuit regions + visualization.
Mechanism to group operators in a circuit in a tree of named regions for
debugging and visualization purposes.

Add `summary()` method to `trait Operator`.  This is so far a very
rudimentary way to obtain operator metadata as a string.  We will
hopefully add a more structured metadata API in the future.

Also implement conversion of a circuit graph to dot format.  This is not
a good long-term debugging solution, as it lacks scalability,
interactivity, and a good layout engine, but it's easy to implement and
is useful for debugging the DBSP library.