[6/15] Z-set revamp: Iterate to a fixed point.
This commit implements the `Circuit::fixedpoint` API and related
infrastructure.  This method creates a child circuit that iterates
until reaching a fixed point, i.e., a state where the outputs of all
operators are guaranteed to remain the same, should the nested clock
continue ticking.

The fixed point check is implemented by checking the following
condition:

* All operators in the circuit are in such a state that, if their inputs
  remain constant (i.e., all future inputs are identical to the last
  input), then their outputs remain constant too.

This is a necessary and sufficient condition that is also easy to check
by asking each operator if it is in a stable state via the new
`Operator::fixedpoint` API.  However, the cost of checking this
condition precisely can be high for some operators.  For instance, delay
operators `Z1` and `Z1Nested) require storing the last two versions of
the state instead of one and comparing them at each cycle.  Such
operators instead implement imprecise conservative checks, e.g., check
for a _specific_ fixed point, e.g., a fixed point where both input and
output of the operator is zero (or empty).  As a result, the circuit may
fail to detect other fixed points and will iterate forever.  The goal is
to evolve the design so that circuits created using the high-level API
(`Stream::xxx` methods) implement accurate fixed point checks.

This commit also adds an unrelated change: the
`StrictOperator::get_final_output` method.  This method extracts the
last output of the operator before the end of the current clock epoch in
order to send it to the parent circuit.  In the past, we simply invoked
the `get_output` method, but that way the operator cannot optimize for
the case when this is the last call before `clock_end`  (e.g., return an
owned value where it would have to clone otherwise).